patient_management/requirements.txt
django==5.1.2
djangorestframework==3.15.2
markdown==3.7
django-filter==24.3